+++ src/c++/perf_analyzer/genai-perf/genai_perf/main.py
@@ -35,8 +35,8 @@ import genai_perf.logging as logging
 from genai_perf import parser
 from genai_perf.exceptions import GenAIPerfException
 from genai_perf.export_data.output_reporter import OutputReporter
+from genai_perf.llm_inputs.inputs_utils import DEFAULT_STARTING_INDEX
 from genai_perf.llm_inputs.llm_inputs import LlmInputs
-from genai_perf.llm_inputs.shared import DEFAULT_STARTING_INDEX
 from genai_perf.plots.plot_config_parser import PlotConfigParser
 from genai_perf.plots.plot_manager import PlotManager
 from genai_perf.profile_data_parser import LLMProfileDataParser, ProfileDataParser
